+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Changelog
 
 -   Fix flakiness in `test_run` in `pyquil/test/test_quantum_computer.py`
     (@appleby, gh-1190).
+-   Fix a bug in QuantumComputer.experiment that resulted in a TypeError being
+    raised when called multiple times on the same experiment when the underlying QAM
+    was a QVM based on a physical device (@appleby, gh-1188).
 
 [v2.18](https://github.com/rigetti/pyquil/compare/v2.17.0...v2.18.0) (March 3, 2020)
 ------------------------------------------------------------------------------------