+++ CHANGELOG.md
@@ -10,7 +10,9 @@ Changelog
 
 -   Add a section to `CONTRIBUTING.md` about publishing packages to conda-forge
     (@appleby, gh-1186).
--   `controlled` modifier now accepts either a Sequence of control qubits or a single control qubit. Previously, only a single control qubit was supported (@adamglos92, gh-1196)
+-   Correctly insert state preparation code in `Experiment`s *before* main program code
+    (@notmgsk, gh-1189).
+-   `controlled` modifier now accepts either a Sequence of control qubits or a single control qubit. Previously, only a single control qubit was supported (@adamglos92, gh-1196).
 
 ### Bugfixes